requirements.txt
web3
rich
aiohttp
qrcode
mnemonic
pycryptodome 